Restruct mod parquet_reader.
